BottomSheet Functionality
The bottomSheet appear and hide animations are almost complete. Content is also now able to be placed inside of the bottomSheet. Two bugs remain. When closing bottom sheet from half opened there is no smoothing when closing. The Second error is that the boolean that hides the bottom sheet so the ui behind it can be interacted with when it is closed is crashing the app when attemping to swipe the bottomsheet closed. Although it is successful when pressing the profile image to close.